requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
decorator==5.2.1
ipykernel==7.1.0
ipython==9.8.0
jupyter_client==8.6.3
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nltk==3.9.2
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
prompt_toolkit==3.0.52
python-dotenv==1.2.1
requests==2.32.5
tqdm==4.67.1
gradio
streamlit==1.52.1
watchdog==6.0.0